Check for document before initializing the Worker
I believe this is related to https://github.com/chrisguttandin/web-audio-beat-detector/issues/285, as it addresses a similar error "TypeError: The "obj" argument must be an instance of Blob. Received an instance of Blob" when trying to import the module using Remix (an SSR-based framework which identifies the workaround for this issue here: https://remix.run/docs/en/v1/guides/constraints#document-guard)

This fixes the issue for me locally, although I can't be sure how this might impact other configurations.